Implemented more of `BotLootCacheService` and related areas

+++ Core/Models/Spt/Bots/BotLootCache.cs
@@ -5,43 +5,43 @@ namespace Core.Models.Spt.Bots;
 public class BotLootCache
 {
     [JsonPropertyName("backpackLoot")]
-    public Dictionary<string, int>? BackpackLoot { get; set; }
+    public Dictionary<string, double>? BackpackLoot { get; set; }
 
     [JsonPropertyName("pocketLoot")]
-    public Dictionary<string, int>? PocketLoot { get; set; }
+    public Dictionary<string, double>? PocketLoot { get; set; }
 
     [JsonPropertyName("vestLoot")]
-    public Dictionary<string, int>? VestLoot { get; set; }
+    public Dictionary<string, double>? VestLoot { get; set; }
 
     [JsonPropertyName("secureLoot")]
-    public Dictionary<string, int>? SecureLoot { get; set; }
+    public Dictionary<string, double>? SecureLoot { get; set; }
 
     [JsonPropertyName("combinedPoolLoot")]
-    public Dictionary<string, int>? CombinedPoolLoot { get; set; }
+    public Dictionary<string, double>? CombinedPoolLoot { get; set; }
 
     [JsonPropertyName("specialItems")]
-    public Dictionary<string, int>? SpecialItems { get; set; }
+    public Dictionary<string, double>? SpecialItems { get; set; }
 
     [JsonPropertyName("healingItems")]
-    public Dictionary<string, int>? HealingItems { get; set; }
+    public Dictionary<string, double>? HealingItems { get; set; }
 
     [JsonPropertyName("drugItems")]
-    public Dictionary<string, int>? DrugItems { get; set; }
+    public Dictionary<string, double>? DrugItems { get; set; }
 
     [JsonPropertyName("foodItems")]
-    public Dictionary<string, int>? FoodItems { get; set; }
+    public Dictionary<string, double>? FoodItems { get; set; }
 
     [JsonPropertyName("drinkItems")]
-    public Dictionary<string, int>? DrinkItems { get; set; }
+    public Dictionary<string, double>? DrinkItems { get; set; }
 
     [JsonPropertyName("currencyItems")]
-    public Dictionary<string, int>? CurrencyItems { get; set; }
+    public Dictionary<string, double>? CurrencyItems { get; set; }
 
     [JsonPropertyName("stimItems")]
-    public Dictionary<string, int>? StimItems { get; set; }
+    public Dictionary<string, double>? StimItems { get; set; }
 
     [JsonPropertyName("grenadeItems")]
-    public Dictionary<string, int>? GrenadeItems { get; set; }
+    public Dictionary<string, double>? GrenadeItems { get; set; }
 }
 
 public class LootCacheType
